refactor(rome_cli): Split CLI tests
Split the command specific CLI tests into their own modules to reduce the size of the `main.rs` file.
