+++ crates/rome_cli/tests/commands/mod.rs
@@ -1,0 +1,3 @@
+mod check;
+mod ci;
+mod format;
